refactor: fix packages organization
