+++ mvc/src/test/java/nextstep/mvc/DispatcherServletTest.java
@@ -8,9 +8,9 @@ import static org.mockito.Mockito.when;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.controller.asis.ForwardController;
-import nextstep.mvc.controller.asis.HandlerExecutionAdapter;
+import nextstep.mvc.controller.tobe.HandlerExecutionAdapter;
 import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
-import nextstep.mvc.controller.tobe.HandlerControllerAdapter;
+import nextstep.mvc.controller.asis.HandlerControllerAdapter;
 import nextstep.mvc.controller.tobe.HandlerExecution;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
